Fixed the overflow flexbox problem
Flexbox causes overflow hidden to not function as expected. This caused the main-nav-item container to be bigger than it should.

To get around this I absolutely positioned the main-nav-list a tag. This also meant that it would affect the pseudo elements that are already absolute positioned. If I gave the li tag a relative position it would it resulted in the preview only applying to width and length of the `li` tag. This why I created a new nav-list -wrapper element to wrap the main-nav-list a tag in.